ProjectTest/requirements.txt
asyncpg
fastapi
passlib[bcrypt]
pydantic[dotenv]
python-jose[criptography]
python-multipart
sqlmodel
uvicorn

# dev
httpx
pytest
pytest-cov
requests
